refactor(checker): extract union restricted-property rule to submodule
Keeps property_checker.rs under the 2000-LOC architecture ceiling (§19) by
moving the union/intersection restricted-property presence rule into
checkers/property_checker/union_restricted_property.rs. Behavior-preserving.
